refactor!: fetcher and seeder based on chan rpc services
Replacing the legacy fetcher and seeder with implementations based on
decomposed common chan_rpc services.

This is an API and protocol breaking change.
