--- stigmerge-peer/src/fetcher2/mod.rs
@@ -1,5 +1,0 @@
-pub mod block_fetcher;
-pub mod fetcher;
-pub mod piece_verifier;
-pub mod seeder;
-pub mod types;
