+++ stigmerge-peer/src/piece_verifier.rs
