+++ stigmerge-peer/src/seeder.rs
@@ -1,317 +1,379 @@
-use std::path::PathBuf;
+use std::{collections::HashMap, path::PathBuf};
 
-use path_absolutize::*;
+use stigmerge_fileindex::{Index, BLOCK_SIZE_BYTES, PIECE_SIZE_BLOCKS};
 use tokio::{
     fs::File,
     io::{AsyncReadExt, AsyncSeekExt},
     select,
-    sync::watch,
+    sync::broadcast,
 };
 use tokio_util::sync::CancellationToken;
-use tracing::{debug, info, instrument, warn};
-use veilid_core::{Target, VeilidAPIError, VeilidUpdate};
+use veilid_core::VeilidUpdate;
 
-use stigmerge_fileindex::{Index, Indexer, BLOCK_SIZE_BYTES, PIECE_SIZE_BLOCKS};
+use crate::types::{PieceState, ShareInfo};
+use crate::{
+    chan_rpc::ChanServer,
+    piece_map::PieceMap,
+    proto::{self, BlockRequest, Decoder},
+    Error, Peer, Result,
+};
 
-use crate::{
-    peer::with_backoff_reset,
-    proto::{BlockRequest, Decoder, Header, Request},
-    reset_with_backoff,
-};
-use crate::{
-    peer::{Peer, TypedKey},
-    Error, Result,
-};
+#[derive(Debug)]
+pub enum Request {
+    HaveMap,
+}
+
+#[derive(Debug)]
+pub enum Response {
+    HaveMap(PieceMap),
+}
 
 pub struct Seeder<P: Peer> {
     peer: P,
-    index: Index,
-    share_key: TypedKey,
-    target: Target,
-    header: Header,
+    ch: ChanServer<Request, Response>,
+    want_index: Index,
+    root: PathBuf,
+    clients: Clients,
+    files: HashMap<u32, File>,
 }
 
 impl<P: Peer> Seeder<P> {
-    pub async fn new(mut peer: P, index: Index) -> Result<Seeder<P>> {
-        let (share_key, target, header) =
-            with_backoff_reset!(peer, peer.announce_index(&index).await)?;
+    pub fn new(
+        peer: P,
+        ch: ChanServer<Request, Response>,
+        share: ShareInfo,
+        clients: Clients,
+    ) -> Result<Self> {
         Ok(Seeder {
             peer,
-            index,
-            share_key,
-            target,
-            header,
+            ch,
+            want_index: share.want_index,
+            root: share.root,
+            clients,
+            files: HashMap::new(),
         })
     }
 
-    #[deprecated = "use new(peer, index) instead"]
-    #[instrument(skip(peer), level = "debug", err)]
-    pub async fn from_file(peer: P, file: &str) -> Result<Seeder<P>> {
-        // Derive root directory
-        let file_path_buf: PathBuf = PathBuf::from(file);
-        let abs_file = file_path_buf.absolutize()?;
-
-        // Build an index of the content to be shared
-        info!(path = format!("{:?}", file_path_buf), "indexing file");
-        let indexer = Indexer::from_file(abs_file.into())
-            .await
-            .map_err(Error::index)?;
-        let index = indexer.index().await.map_err(Error::index)?;
-        Self::new(peer, index).await
-    }
-
-    pub fn share_key(&self) -> String {
-        self.share_key.to_string()
-    }
-
-    pub fn digest(&self) -> String {
-        hex::encode(self.index.payload().digest())
-    }
-
-    pub async fn seed(mut self, cancel: CancellationToken) -> Result<()> {
-        if self.index.files().len() > 1 {
-            todo!("multi-file seeding not yet supported, sorry!");
-        }
-        let local_single_file = self.index.root().join(self.index.files()[0].path());
-        info!(
-            share_key = format!("{}", self.share_key),
-            file = format!("{:?}", local_single_file),
-            "seeding"
-        );
-
-        let mut fh: File = File::open(local_single_file).await?;
+    pub async fn run(&mut self, cancel: CancellationToken) -> Result<()> {
+        let mut piece_map = PieceMap::new();
         let mut buf = [0u8; BLOCK_SIZE_BYTES];
-
-        let mut updates = self.peer.subscribe_veilid_update();
-
-        let (update_err_tx, mut update_err_rx) = watch::channel(Ok(()));
-
         loop {
             select! {
-                recv_update = updates.recv() => {
-                    let update = recv_update.map_err(Error::other)?;
-                    if let Err(e) = self.handle_update(&mut fh, &mut buf, &update).await {
-                        if let Error::RemoteProtocol(_) = e {
-                            // ignore bad requests
-                        }
-                        else if e.is_resetable() {
-                            update_err_tx.send(Err(e)).map_err(Error::other)?;
-                        }
-                        else if e.is_shutdown() {
-                            return Err(e);
-                        }
-                    }
-                }
-                _ = update_err_rx.changed() => {
-                    // Reset the peer
-                    reset_with_backoff(&mut self.peer, &cancel).await?;
-
-                    // Attempt to reannounce the route
-                    if let Err(e) = self.reannounce_route().await {
-                        warn!(err = format!("{}", e), "failed to reannounce route");
-                        update_err_tx.send(Err(e)).map_err(Error::other)?;
-                    }
-                }
                 _ = cancel.cancelled() => {
-                    info!("seeding cancelled");
-                    break
-                }
-            }
-        }
-        Ok(())
-    }
-
-    async fn handle_update(
-        &mut self,
-        fh: &mut File,
-        buf: &mut [u8],
-        update: &VeilidUpdate,
-    ) -> Result<()> {
-        match update {
-            &VeilidUpdate::AppCall(ref app_call) => {
-                let block_request =
-                    match Request::decode(app_call.message()).map_err(Error::remote_protocol)? {
-                        Request::BlockRequest(req) => req,
-                        _ => return Ok(()),
-                    };
-                // TODO: mmap would enable more concurrency here, but might not be as cross-platform?
-                fh.seek(std::io::SeekFrom::Start(
-                    // TODO: wire this through Index to support multifile
-                    ((block_request.piece as usize * PIECE_SIZE_BLOCKS * BLOCK_SIZE_BYTES)
-                        + (block_request.block as usize * BLOCK_SIZE_BYTES))
-                        as u64,
-                ))
-                .await?;
-                let rd = fh.read(buf).await?;
-                // TODO: Don't block here; we could handle another request in the meantime
-                self.peer
-                    .reply_block_contents(app_call.id(), &buf[0..rd])
-                    .await?;
-                Ok(())
-            }
-            &VeilidUpdate::RouteChange(ref route_change) => {
-                let target_route_id = match self.target {
-                    Target::NodeId(_) => return Ok(()),
-                    Target::PrivateRoute(ref route_id) => route_id.to_owned(),
-                };
-                if !route_change.dead_routes.contains(&target_route_id) {
                     return Ok(());
                 }
-                debug!(target = target_route_id.to_string(), "route changed");
-
-                self.reannounce_route().await?;
-
-                Ok(())
+                res = self.ch.rx.recv() => {
+                    if let Some(Request::HaveMap) = res {
+                        self.ch.tx.send(Response::HaveMap(piece_map.clone())).await.map_err(Error::other)?;
+                    }
+                }
+                res = self.clients.verified_rx.recv() => {
+                    let piece_state = res.map_err(Error::other)?;
+                    piece_map.set(piece_state.piece_index.try_into().unwrap());
+                }
+                res = self.clients.update_rx.recv() => {
+                    let update = res.map_err(Error::other)?;
+                    match update {
+                        VeilidUpdate::AppCall(veilid_app_call) => {
+                            let req = proto::Request::decode(veilid_app_call.message()).map_err(Error::RemoteProtocol)?;
+                            match req {
+                                proto::Request::BlockRequest(block_req) => {
+                                    if piece_map.get(block_req.piece) {
+                                        let rd = self.read_block_into(&block_req, &mut buf).await?;
+                                        self.peer.reply_block_contents(veilid_app_call.id(), &buf[..rd]).await?;
+                                    } else {
+                                        self.peer.reply_block_contents(veilid_app_call.id(), &[]).await?;
+                                    }
+                                }
+                                _ => {}  // Ignore other request types
+                            }
+                        }
+                        _ => {}
+                    }
+                }
             }
-            &VeilidUpdate::Shutdown => Err(Error::Fault(crate::error::Unexpected::Veilid(
-                VeilidAPIError::Shutdown,
-            ))),
-            _ => Ok(()),
         }
     }
 
-    async fn reannounce_route(&mut self) -> Result<()> {
-        let (target, header) = with_backoff_reset!(
-            self.peer,
-            self.peer
-                .reannounce_route(
-                    &self.share_key,
-                    Some(self.target.to_owned()),
-                    &self.index,
-                    &self.header,
-                )
-                .await
-        )?;
-        self.target = target;
-        self.header = header;
-        Ok(())
+    async fn read_block_into(&mut self, block_req: &BlockRequest, buf: &mut [u8]) -> Result<usize> {
+        let fh = self.get_file_for_block(&block_req).await?;
+        fh.seek(std::io::SeekFrom::Start(
+            ((TryInto::<usize>::try_into(block_req.piece).unwrap()
+                * PIECE_SIZE_BLOCKS
+                * BLOCK_SIZE_BYTES)
+                + (TryInto::<usize>::try_into(block_req.block).unwrap() * BLOCK_SIZE_BYTES))
+                .try_into()
+                .unwrap(),
+        ))
+        .await?;
+        let rd = fh.read(buf).await?;
+        Ok(rd)
     }
+
+    async fn get_file_for_block(&mut self, block_req: &BlockRequest) -> Result<&mut File> {
+        if !self.files.contains_key(&block_req.piece) {
+            // FIXME: does not support multifile; piece -> file mapping should be handled by Index
+            let file_path = self.root.join(self.want_index.files()[0].path());
+            let fh = File::open(file_path).await?;
+            self.files.insert(block_req.piece, fh);
+        }
+        Ok(self.files.get_mut(&block_req.piece).unwrap())
+    }
+}
+
+pub struct Clients {
+    pub verified_rx: broadcast::Receiver<PieceState>,
+    pub update_rx: broadcast::Receiver<VeilidUpdate>,
 }
 
 #[cfg(test)]
 mod tests {
     use std::{
-        str::FromStr,
         sync::{Arc, Mutex},
         time::Duration,
     };
 
-    use tokio::time::sleep;
-    use veilid_core::{
-        OperationId, TimestampDuration, TypedKey, VeilidAppCall, VeilidRouteChange,
-        VeilidStateAttachment, VeilidUpdate,
-    };
+    use tokio::{sync::mpsc, time};
+    use tokio_util::sync::CancellationToken;
+    use veilid_core::{OperationId, VeilidAppCall};
 
     use crate::{
-        proto::{BlockRequest, Encoder, Request},
+        chan_rpc::pipe,
+        proto::{BlockRequest, Encoder, Header},
         tests::{temp_file, StubPeer},
-        Observable,
     };
 
     use super::*;
 
+    fn test_header() -> Header {
+        Header::new([0xab; 32], 42, 1, [0xcd; 99].as_slice(), None, None)
+    }
+
     #[tokio::test]
-    async fn from_dht_ok() {
-        // Temp file to seed, index it, create a stub key for target and share
-        let tf = temp_file(0xa5u8, 1048576);
-        let tf_path = std::env::temp_dir().join(tf.path()).to_owned();
-        let announce_indexer = Indexer::from_file(tf_path.clone().into())
-            .await
-            .expect("index");
-        let announce_index = announce_indexer.index().await.expect("index");
-        let key = TypedKey::from_str("VLD0:cCHB85pEaV4bvRfywxnd2fRNBScR64UaJC8hoKzyr3M").unwrap();
-        let key_internal = key.clone();
+    async fn test_seeder_block_request_for_verified_piece() {
+        // Create a test file with known content
+        const BLOCK_DATA: u8 = 0xa5;
+        let tf = temp_file(BLOCK_DATA, BLOCK_SIZE_BYTES * PIECE_SIZE_BLOCKS * 2); // 2 pieces
+        let tf_path = tf.path().to_path_buf();
+        let root_dir = tf_path.parent().unwrap().to_path_buf();
 
-        // Track reannounce call
-        let reannounce_index = announce_index.clone();
-        let reannounce_calls = Arc::new(Mutex::new(0));
-        let reannounce_calls_internal = reannounce_calls.clone();
+        // Create a mock index
+        let index = create_test_index(&tf_path).await;
 
-        // Track reply_block_contents
-        let reply_calls = Arc::new(Mutex::new(0));
-        let reply_calls_internal = reply_calls.clone();
+        // Set up channels
+        let (verified_tx, verified_rx) = broadcast::channel(16);
 
-        let mut stub_peer = StubPeer::new();
-        let update_tx = stub_peer.update_tx.clone();
-        stub_peer.reset_result = Arc::new(Mutex::new(move || Ok(())));
-        stub_peer.announce_result = Arc::new(Mutex::new(move |_index: &Index| {
-            let index_bytes = announce_index.encode().expect("encode index");
-            Ok((
-                key_internal.clone(),
-                Target::PrivateRoute(key_internal.value.clone()),
-                Header::from_index(&announce_index, &index_bytes, &[0xde, 0xad, 0xbe, 0xef]),
-            ))
-        }));
-        stub_peer.reannounce_route_result = Arc::new(Mutex::new(
-            move |_key: &TypedKey, _target: Option<Target>, _index: &Index, _header: &Header| {
-                let index_bytes = reannounce_index.encode().expect("encode index");
-                (*(reannounce_calls_internal.lock().unwrap())) += 1;
-                Ok((
-                    Target::PrivateRoute(key_internal.value.clone()),
-                    Header::from_index(&reannounce_index, &index_bytes, &[0xde, 0xad, 0xbe, 0xef]),
-                ))
-            },
-        ));
-        stub_peer.reply_block_contents_result =
-            Arc::new(Mutex::new(move |_id: OperationId, _contents: &[u8]| {
-                (*(reply_calls_internal.lock().unwrap())) += 1;
+        // Create a stub peer with mock reply_block_contents
+        let mut peer = StubPeer::new();
+        let update_tx = peer.update_tx.clone();
+        let reply_contents_called = Arc::new(Mutex::new(false));
+        let reply_contents_data = Arc::new(Mutex::new(Vec::new()));
+        let reply_contents_called_clone = reply_contents_called.clone();
+        let reply_contents_data_clone = reply_contents_data.clone();
+
+        let (replied_tx, mut replied_rx) = mpsc::channel(1);
+
+        peer.reply_block_contents_result =
+            Arc::new(Mutex::new(move |_call_id: OperationId, contents: &[u8]| {
+                *reply_contents_called_clone.lock().unwrap() = true;
+                reply_contents_data_clone
+                    .lock()
+                    .unwrap()
+                    .extend_from_slice(contents);
+                replied_tx.try_send(()).expect("replied");
                 Ok(())
             }));
-        let rp = Observable::new(stub_peer);
 
-        // Simulate getting connected to network, normally track_node_state
-        // would set this when the node comes online.
-        update_tx
-            .send(VeilidUpdate::Attachment(Box::new(VeilidStateAttachment {
-                state: veilid_core::AttachmentState::AttachedGood,
-                public_internet_ready: true,
-                local_network_ready: true,
-                attached_uptime: None,
-                uptime: TimestampDuration::new(0),
-            })))
-            .expect("send veilid update");
+        let (mut client_ch, server_ch) = pipe(1);
 
-        let indexer = Indexer::from_file(tf_path.into()).await.expect("indexer");
-        let index = indexer.index().await.expect("index");
-        let seeder = Seeder::new(rp, index).await.expect("new seeder");
+        // Create share info
+        let share_info = ShareInfo {
+            header: test_header(),
+            want_index: index,
+            root: root_dir,
+        };
+
+        // Create clients
+        let clients = Clients {
+            verified_rx,
+            update_rx: peer.update_tx.subscribe(),
+        };
+
+        // Create cancellation token
         let cancel = CancellationToken::new();
-        let update_tx_internal = update_tx.clone();
-        tokio::spawn(async move {
-            // Simulate a request for a block
-            let request_bytes = Request::BlockRequest(BlockRequest { piece: 0, block: 0 })
-                .encode()
-                .expect("encode block request");
-            sleep(Duration::from_millis(50)).await;
-            update_tx_internal
-                .send(VeilidUpdate::AppCall(Box::new(VeilidAppCall::new(
-                    None,
-                    None,
-                    request_bytes,
-                    OperationId::new(42u64),
-                ))))
-                .expect("send app call");
+        let cancel_clone = cancel.clone();
 
-            // Create a route change
-            sleep(Duration::from_millis(50)).await;
-            update_tx_internal
-                .send(VeilidUpdate::RouteChange(Box::new(VeilidRouteChange {
-                    dead_routes: vec![key_internal.value],
-                    dead_remote_routes: vec![],
-                })))
-                .expect("send route change");
-
-            // Shut down the node
-            sleep(Duration::from_millis(50)).await;
-            update_tx_internal
-                .send(VeilidUpdate::Shutdown)
-                .expect("shutdown");
-
-            Ok::<(), Error>(())
+        // Run seeder in a task
+        let seeder_task = tokio::spawn(async move {
+            let mut seeder = Seeder::new(peer, server_ch, share_info, clients)?;
+            seeder.run(cancel_clone).await
         });
-        let result = seeder.seed(cancel).await;
-        assert!(matches!(result, Err(_)));
 
-        assert_eq!(*(reannounce_calls.lock().unwrap()), 1);
-        assert_eq!(*(reply_calls.lock().unwrap()), 1);
+        // First, send a verified piece notification with confirmation it's applied
+        // That's necessary because tokio::select! future selection is non-deterministic
+        let mut confirmed_verify = false;
+        for _ in 0..10 {
+            let piece_state = PieceState::new(0, 0, 0, PIECE_SIZE_BLOCKS, PIECE_SIZE_BLOCKS - 1);
+            verified_tx.send(piece_state).expect("send verified piece");
+
+            client_ch
+                .tx
+                .send(Request::HaveMap)
+                .await
+                .expect("requested havemap");
+            let Response::HaveMap(have_map) = client_ch.rx.recv().await.expect("havemap response");
+            if !have_map.is_empty() {
+                confirmed_verify = true;
+                break;
+            }
+        }
+        assert!(confirmed_verify, "confirm verified block");
+
+        // Send a block request for the verified piece
+        let block_req = BlockRequest { piece: 0, block: 0 };
+        let req = proto::Request::BlockRequest(block_req);
+        let encoded_req = req.encode().expect("encode request");
+        let app_call = VeilidAppCall::new(None, None, encoded_req, 42u64.into());
+        update_tx
+            .send(VeilidUpdate::AppCall(Box::new(app_call)))
+            .expect("send app call");
+
+        time::timeout(Duration::from_secs(1), replied_rx.recv())
+            .await
+            .expect("replied");
+
+        // Cancel the seeder
+        cancel.cancel();
+
+        // Verify seeder exits cleanly
+        seeder_task.await.expect("seeder task").expect("seeder run");
+
+        // Verify reply_block_contents was called
+        assert!(
+            *reply_contents_called.lock().unwrap(),
+            "reply_block_contents should be called"
+        );
+
+        // Verify the data returned matches what we expect
+        let reply_data = reply_contents_data.lock().unwrap();
+        assert_eq!(
+            reply_data.len(),
+            BLOCK_SIZE_BYTES,
+            "should return full block"
+        );
+        assert!(
+            reply_data.iter().all(|&b| b == BLOCK_DATA),
+            "all bytes should match the pattern"
+        );
+    }
+
+    #[tokio::test]
+    async fn test_seeder_block_request_for_unverified_piece() {
+        // Create a test file with known content
+        const BLOCK_DATA: u8 = 0xa5;
+        let tf = temp_file(BLOCK_DATA, BLOCK_SIZE_BYTES * PIECE_SIZE_BLOCKS * 2); // 2 pieces
+        let tf_path = tf.path().to_path_buf();
+        let root_dir = tf_path.parent().unwrap().to_path_buf();
+
+        // Create a mock index
+        let index = create_test_index(&tf_path).await;
+
+        // Set up channels
+        let (_verified_tx, verified_rx) = broadcast::channel(16);
+
+        // Create a stub peer with mock reply_block_contents
+        let mut peer = StubPeer::new();
+        let update_tx = peer.update_tx.clone();
+
+        let reply_contents_called = Arc::new(Mutex::new(false));
+        let reply_contents_data = Arc::new(Mutex::new(Vec::new()));
+        let reply_contents_called_clone = reply_contents_called.clone();
+        let reply_contents_data_clone = reply_contents_data.clone();
+
+        let (replied_tx, mut replied_rx) = mpsc::channel(1);
+
+        peer.reply_block_contents_result =
+            Arc::new(Mutex::new(move |_call_id: OperationId, contents: &[u8]| {
+                *reply_contents_called_clone.lock().unwrap() = true;
+                reply_contents_data_clone
+                    .lock()
+                    .unwrap()
+                    .extend_from_slice(contents);
+                replied_tx.try_send(()).expect("replied");
+                Ok(())
+            }));
+
+        let (_client_ch, server_ch) = pipe(1);
+
+        // Create share info
+        let share_info = ShareInfo {
+            header: test_header(),
+            want_index: index,
+            root: root_dir,
+        };
+
+        // Create clients
+        let clients = Clients {
+            verified_rx,
+            update_rx: peer.update_tx.subscribe(),
+        };
+
+        // Create seeder
+        let mut seeder = Seeder::new(peer, server_ch, share_info, clients).expect("create seeder");
+
+        // Create cancellation token
+        let cancel = CancellationToken::new();
+        let cancel_clone = cancel.clone();
+
+        // Run seeder in a task
+        let seeder_task = tokio::spawn(async move { seeder.run(cancel_clone).await });
+
+        // Send a block request for an unverified piece
+        let block_req = BlockRequest { piece: 0, block: 0 };
+        let req = proto::Request::BlockRequest(block_req);
+        let encoded_req = req.encode().expect("encode request");
+
+        let app_call = VeilidAppCall::new(None, None, encoded_req, 42u64.into());
+
+        update_tx
+            .send(VeilidUpdate::AppCall(Box::new(app_call)))
+            .expect("send app call");
+
+        time::timeout(Duration::from_secs(1), replied_rx.recv())
+            .await
+            .expect("replied");
+
+        // Cancel the seeder
+        cancel.cancel();
+
+        // Verify seeder exits cleanly
+        let result = seeder_task.await.expect("seeder task");
+        assert!(
+            result.is_ok(),
+            "seeder should exit cleanly after handling block request"
+        );
+
+        // Verify reply_block_contents was called
+        assert!(
+            *reply_contents_called.lock().unwrap(),
+            "reply_block_contents should be called"
+        );
+
+        // Verify empty data was returned for unverified piece
+        let reply_data = reply_contents_data.lock().unwrap();
+        assert_eq!(
+            reply_data.len(),
+            0,
+            "should return empty data for unverified piece"
+        );
+    }
+
+    // Helper function to create a test index
+    async fn create_test_index(file_path: &PathBuf) -> Index {
+        // Create index from the file using Indexer
+        let indexer = stigmerge_fileindex::Indexer::from_file(file_path.clone())
+            .await
+            .expect("create indexer");
+        indexer.index().await.expect("create index")
     }
 }
