+++ stigmerge-peer/src/types.rs
